practical_projects/Twilio_FIFA_bot/requirements.txt
Flask==1.1.2
python-dateutil==2.8.1
requests==2.24.0
twilio==6.45.2
